refactor(fieldmap): extract north arrow into separate component
